+++ apps/native-component-list/src/screens/UI/JetpackComposePrimitivesScreen.android.tsx
@@ -1,14 +1,32 @@
-import { Container, Column, Row, Text } from '@expo/ui/jetpack-compose-primitives';
+import {
+  background,
+  size,
+  weight,
+  testID,
+  matchParentSize,
+  blur,
+  clickable,
+  animateContentSize,
+  paddingAll,
+  Host,
+  Column,
+  Row,
+  Box,
+  Text,
+} from '@expo/ui/jetpack-compose';
 import { useState } from 'react';
 import { Button, View, StyleSheet, Text as RNText } from 'react-native';
 
 export default function JetpackComposePrimitivesScreen() {
   const [checked, setChecked] = useState<boolean>(true);
   return (
-    <Container style={styles.container} testID="container">
+    <Host style={styles.container} modifiers={[testID('container')]}>
       <Column verticalArrangement="spaceEvenly" horizontalAlignment="center">
         {/* Example 1: Row with Text and Switch */}
-        <Row horizontalArrangement="spaceBetween" verticalAlignment="center" testID="leftTextRow">
+        <Row
+          horizontalArrangement="spaceBetween"
+          verticalAlignment="center"
+          modifiers={[testID('leftTextRow')]}>
           <Text fontSize={18} fontWeight="bold" color="#ff0000">
             Left Text
           </Text>
@@ -16,13 +34,25 @@ export default function JetpackComposePrimitivesScreen() {
 
         {/* Example 2: Column with different Text styles */}
         <Column verticalArrangement="spaceEvenly" horizontalAlignment="center">
-          <Text fontSize={24} fontWeight="bold" color="#ff0000" testID="largeBoldText">
+          <Text
+            fontSize={24}
+            fontWeight="bold"
+            color="#ff0000"
+            modifiers={[testID('largeBoldText')]}>
             Large Bold Text
           </Text>
-          <Text fontSize={16} fontWeight="normal" color="#00ff00" testID="mediumNormalText">
+          <Text
+            fontSize={16}
+            fontWeight="normal"
+            color="#00ff00"
+            modifiers={[testID('mediumNormalText')]}>
             Medium Normal Text
           </Text>
-          <Text fontSize={12} fontWeight="300" color="#666666" testID="smallLightText">
+          <Text
+            fontSize={12}
+            fontWeight="300"
+            color="#666666"
+            modifiers={[testID('smallLightText')]}>
             Small Light Text
           </Text>
           <View style={{ width: 200, height: 100, backgroundColor: 'green' }}>
@@ -32,26 +62,68 @@ export default function JetpackComposePrimitivesScreen() {
         </Column>
 
         {/* Example 3: Nested Row and Column with Text */}
-        <Row horizontalArrangement="spaceAround" verticalAlignment="center" testID="rowParent">
-          <Column verticalArrangement="center" horizontalAlignment="center" testID="nestedColumn1">
-            <Text fontSize={16} fontWeight="500" testID="nestedColumn1Text">
+        <Row
+          horizontalArrangement="spaceAround"
+          verticalAlignment="center"
+          modifiers={[size(300, 300), background('#ffdddd'), testID('rowParent')]}>
+          <Column
+            modifiers={[background('#ff0000'), weight(2), testID('nestedColumn1')]}
+            verticalArrangement="center"
+            horizontalAlignment="center">
+            <Text fontSize={16} fontWeight="500" modifiers={[testID('nestedColumn1Text')]}>
               Nested
             </Text>
             <Text fontSize={14} color="#333333">
               Column 1
             </Text>
           </Column>
-          <Column verticalArrangement="center" horizontalAlignment="center" testID="nestedColumn2">
-            <Text fontSize={16} fontWeight="500" testID="nestedColumn2Text">
+          <Column
+            verticalArrangement="center"
+            horizontalAlignment="center"
+            modifiers={[testID('nestedColumn2')]}>
+            <Text fontSize={16} fontWeight="500" modifiers={[testID('nestedColumn2Text')]}>
               Nested
             </Text>
             <Text fontSize={14} color="#333333">
               Column 2
             </Text>
           </Column>
+          <Column
+            modifiers={[background('#ff0000'), weight(4), testID('nestedColumn3')]}
+            verticalArrangement="center"
+            horizontalAlignment="center">
+            <Text fontSize={16} fontWeight="500" modifiers={[testID('nestedColumn2Text')]}>
+              Nested
+            </Text>
+            <Text fontSize={14} color="#333333">
+              Column 3
+            </Text>
+          </Column>
+          <Column
+            verticalArrangement="center"
+            horizontalAlignment="center"
+            modifiers={[testID('nestedColumn4')]}>
+            <Text fontSize={16} fontWeight="500" modifiers={[testID('nestedColumn4text')]}>
+              Nested
+            </Text>
+            <Text fontSize={14} color="#333333">
+              Column 4
+            </Text>
+          </Column>
         </Row>
+        {/* Example 3: Two nested boxes */}
+        <Box
+          modifiers={[
+            blur(10),
+            size(200, checked ? 100 : 200),
+            animateContentSize(0.5, 100),
+            background('#ffdddd'),
+            clickable(() => setChecked((c) => !c)),
+          ]}>
+          <Box modifiers={[matchParentSize(), background('#ddddff'), paddingAll(30)]} />
+        </Box>
       </Column>
-    </Container>
+    </Host>
   );
 }