+++ packages/@expo/cli/src/start/server/metro/createServerRouteMiddleware.ts
@@ -41,6 +41,7 @@ export function createRouteHandlerMiddleware(
       functionFilePath: string
     ) => Promise<null | Record<string, Function> | Response>;
     config: ProjectConfig;
+    headers: Record<string, string | string[]>;
   } & import('expo-router/build/routes-manifest').Options
 ) {
   if (!resolveFrom.silent(projectRoot, 'expo-router')) {