+++ packages/expo-dev-launcher/android/src/debug/java/expo/modules/devlauncher/launcher/loaders/DevLauncherPublishedAppLoader.kt
@@ -1,7 +1,7 @@
 package expo.modules.devlauncher.launcher.loaders
 
 import android.content.Context
-import expo.interfaces.devmenu.ReactHostWrapper
+import com.facebook.react.ReactHost
 import expo.modules.devlauncher.helpers.injectLocalBundleLoader
 import expo.modules.devlauncher.launcher.DevLauncherControllerInterface
 import expo.modules.manifests.core.Manifest
@@ -13,7 +13,7 @@ import expo.modules.manifests.core.Manifest
 class DevLauncherPublishedAppLoader(
   manifest: Manifest,
   private val localBundlePath: String,
-  private val appHost: ReactHostWrapper,
+  private val appHost: ReactHost,
   context: Context,
   controller: DevLauncherControllerInterface
 ) : DevLauncherExpoAppLoader(manifest, appHost, context, controller) {