+++ packages/expo-dev-launcher/android/src/debug/java/expo/modules/devlauncher/launcher/loaders/DevLauncherReactNativeAppLoader.kt
@@ -2,13 +2,13 @@ package expo.modules.devlauncher.launcher.loaders
 
 import android.content.Context
 import android.net.Uri
-import expo.interfaces.devmenu.ReactHostWrapper
+import com.facebook.react.ReactHost
 import expo.modules.devlauncher.helpers.injectReactInterceptor
 import expo.modules.devlauncher.launcher.DevLauncherControllerInterface
 
 class DevLauncherReactNativeAppLoader(
   private val url: Uri,
-  private val appHost: ReactHostWrapper,
+  private val appHost: ReactHost,
   private val context: Context,
   controller: DevLauncherControllerInterface
 ) : DevLauncherAppLoader(appHost, context, controller) {