+++ packages/expo-ui/android/src/main/java/expo/modules/ui/TextInputView.kt
@@ -22,6 +22,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.KeyboardCapitalization
 import expo.modules.kotlin.views.AutoSizingComposable
 import expo.modules.kotlin.views.Direction
+import expo.modules.kotlin.views.ComposableScope
 import java.util.EnumSet
 
 data class TextInputProps(
@@ -62,7 +63,7 @@ private fun String.autoCapitalize(): KeyboardCapitalization {
 }
 
 class TextInputView(context: Context, appContext: AppContext) :
-  ExpoComposeView<TextInputProps>(context, appContext, withHostingView = true) {
+  ExpoComposeView<TextInputProps>(context, appContext) {
   override val props = TextInputProps()
   private val onValueChanged by EventDispatcher()
 
@@ -76,7 +77,7 @@ class TextInputView(context: Context, appContext: AppContext) :
     }
 
   @Composable
-  override fun Content(modifier: Modifier) {
+  override fun ComposableScope.Content() {
     AutoSizingComposable(shadowNodeProxy, axis = EnumSet.of(Direction.VERTICAL)) {
       TextField(
         value = requireNotNull(textState.value),