+++ packages/expo-ui/android/src/main/java/expo/modules/ui/button/Button.kt
@@ -19,6 +19,7 @@ import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.unit.dp
 import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.records.Field
@@ -26,11 +27,14 @@ import expo.modules.kotlin.records.Record
 import expo.modules.kotlin.views.ComposeProps
 import java.io.Serializable
 import expo.modules.kotlin.types.Enumerable
+import expo.modules.kotlin.views.ComposableScope
 import expo.modules.ui.DynamicTheme
 import expo.modules.ui.ExpoModifier
+import expo.modules.ui.ShapeRecord
 import expo.modules.ui.compose
 import expo.modules.ui.fromExpoModifiers
 import expo.modules.ui.getImageVector
+import expo.modules.ui.shapeFromShapeRecord
 
 open class ButtonPressedEvent() : Record, Serializable
 
@@ -62,12 +66,21 @@ data class ButtonProps(
   val elementColors: MutableState<ButtonColors> = mutableStateOf(ButtonColors()),
   val leadingIcon: MutableState<String?> = mutableStateOf(null),
   val trailingIcon: MutableState<String?> = mutableStateOf(null),
-  val disabled: MutableState<Boolean> = mutableStateOf(false),
-  val modifiers: MutableState<List<ExpoModifier>> = mutableStateOf(emptyList())
+  val disabled: MutableState<Boolean?> = mutableStateOf(false),
+  val modifiers: MutableState<List<ExpoModifier>?> = mutableStateOf(emptyList()),
+  val shape: MutableState<ShapeRecord?> = mutableStateOf(null)
 ) : ComposeProps
 
 @Composable
-fun StyledButton(variant: ButtonVariant, colors: ButtonColors, disabled: Boolean, onPress: () -> Unit, modifier: Modifier = Modifier, content: @Composable (RowScope.() -> Unit)) {
+fun StyledButton(
+  variant: ButtonVariant,
+  colors: ButtonColors,
+  disabled: Boolean,
+  onPress: () -> Unit,
+  modifier: Modifier = Modifier,
+  shape: Shape?,
+  content: @Composable (RowScope.() -> Unit)
+) {
   when (variant) {
     ButtonVariant.BORDERED -> FilledTonalButton(
       onPress,
@@ -79,6 +92,7 @@ fun StyledButton(variant: ButtonVariant, colors: ButtonColors, disabled: Boolean
         disabledContainerColor = colors.disabledContainerColor.compose,
         disabledContentColor = colors.disabledContentColor.compose
       ),
+      shape = shape ?: ButtonDefaults.filledTonalShape,
       modifier = modifier
     )
 
@@ -92,6 +106,7 @@ fun StyledButton(variant: ButtonVariant, colors: ButtonColors, disabled: Boolean
         disabledContainerColor = colors.disabledContainerColor.compose,
         disabledContentColor = colors.disabledContentColor.compose
       ),
+      shape = shape ?: ButtonDefaults.textShape,
       modifier = modifier
     )
 
@@ -105,6 +120,7 @@ fun StyledButton(variant: ButtonVariant, colors: ButtonColors, disabled: Boolean
         disabledContainerColor = colors.disabledContainerColor.compose,
         disabledContentColor = colors.disabledContentColor.compose
       ),
+      shape = shape ?: ButtonDefaults.outlinedShape,
       modifier = modifier
     )
 
@@ -118,6 +134,7 @@ fun StyledButton(variant: ButtonVariant, colors: ButtonColors, disabled: Boolean
         disabledContainerColor = colors.disabledContainerColor.compose,
         disabledContentColor = colors.disabledContentColor.compose
       ),
+      shape = shape ?: ButtonDefaults.elevatedShape,
       modifier = modifier
     )
 
@@ -131,13 +148,14 @@ fun StyledButton(variant: ButtonVariant, colors: ButtonColors, disabled: Boolean
         disabledContainerColor = colors.disabledContainerColor.compose,
         disabledContentColor = colors.disabledContentColor.compose
       ),
+      shape = shape ?: ButtonDefaults.shape,
       modifier = modifier
     )
   }
 }
 
 class Button(context: Context, appContext: AppContext) :
-  ExpoComposeView<ButtonProps>(context, appContext, withHostingView = true) {
+  ExpoComposeView<ButtonProps>(context, appContext) {
   override val props = ButtonProps()
   private val onButtonPressed by EventDispatcher<ButtonPressedEvent>()
 
@@ -147,22 +165,25 @@ class Button(context: Context, appContext: AppContext) :
   }
 
   @Composable
-  override fun Content(modifier: Modifier) {
+  override fun ComposableScope.Content() {
     val (variant) = props.variant
     val (text) = props.text
     val (colors) = props.elementColors
     val (leadingIcon) = props.leadingIcon
     val (trailingIcon) = props.trailingIcon
     val (disabled) = props.disabled
+
     DynamicTheme {
       StyledButton(
         variant ?: ButtonVariant.DEFAULT,
         colors,
-        disabled,
+        disabled ?: false,
         onPress = { onButtonPressed.invoke(ButtonPressedEvent()) },
-        modifier = Modifier.fromExpoModifiers(props.modifiers.value)
+        modifier = Modifier.fromExpoModifiers(props.modifiers.value),
+        shape = shapeFromShapeRecord(props.shape.value)
       ) {
         Row(verticalAlignment = Alignment.CenterVertically) {
+          Children(ComposableScope(rowScope = this))
           leadingIcon?.let { iconName ->
             getImageVector(iconName)?.let {
               Icon(