+++ packages/expo-ui/ios/ExpoUIModule.swift
@@ -22,6 +22,7 @@ public final class ExpoUIModule: Module {
     View(ExpoUI.ContextMenu.self)
     View(ExpoUI.ContextMenuActivationElement.self)
     View(ExpoUI.ContextMenuPreview.self)
+    View(ExpoUI.ContextMenuContent.self)
     View(FormView.self)
     View(GaugeView.self)
     View(GroupView.self)
@@ -73,5 +74,6 @@ public final class ExpoUIModule: Module {
     View(CapsuleView.self)
     View(CircleView.self)
     View(ConcentricRectangleView.self)
+    View(DividerView.self)
   }
 }